# OnlineCinema/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Фильмотека",
        "title": "Фильмотека API",
        "contact": {},
        "version": "1.0"
    },
    "host": ":8080",
    "paths": {
        "/CreateActor": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Функция добавляет нового актёра",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Добавить актёра",
                "operationId": "CreateActor",
                "parameters": [
                    {
                        "description": "Информация об актёре",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Actor"
                        }
                    }
                }
            }
        },
...
        "/DeleteActor": {
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Удаляет актёра",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Удалить актёра",
                "operationId": "DeleteActor",
                "parameters": [
                    {
                        "description": "Информация об актёре",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/GetActors": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Возвращает всех актёров фильмотеки",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Показать актёров",
                "operationId": "GetActors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.OutActor"
                        }
                    }
                }
            }
        },
...
        "/UpdateActor": {
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "обновляет данные об актере",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "actors"
                ],
                "summary": "Изменить актёра",
                "operationId": "UpdateActor",
                "parameters": [
                    {
                        "description": "Информация об актёре",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.Actor"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "storage.Actor": {
            "type": "object",
            "properties": {
                "DeleteFilms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "actorname": {
                    "type": "string"
                },
                "addfilms": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "birthdate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
        "storage.Film": {
            "type": "object",
            "properties": {
                "addActors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "deleteActors": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "description": {
                    "type": "string"
                },
                "filmname": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "string"
                },
                "releasedate": {
                    "type": "string"
                }
            }
        },
...
        "storage.OutActor": {
            "type": "object",
            "properties": {
                "actorname": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "string"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Film"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}